% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-app.R
\name{analytics_app}
\alias{analytics_app}
\title{Run example telemetry analytics dashboard}
\usage{
analytics_app(data_storage, run_app = FALSE)
}
\arguments{
\item{data_storage}{data_storage instance that will handle all backend read
and writes.}

\item{run_app}{flag that indicates if shinyApp should be called in the
function or return the UI and Server instead.}
}
\description{
Run example telemetry analytics dashboard
}
