% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-panel-top.R
\name{analytics_server}
\alias{analytics_server}
\title{Stats Shiny server}
\usage{
analytics_server(
  get_user,
  allow_admin_rule = function(session) TRUE,
  db_credentials
)
}
\arguments{
\item{get_user}{function to retrieve user name}

\item{allow_admin_rule}{function that asserts if user is administrator}

\item{db_credentials}{data.frame with database config parameters}
}
\description{
Stats Shiny server
}
