% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{AVAILABLE_DB_DRIVERS}
\alias{AVAILABLE_DB_DRIVERS}
\title{Possible drivers allowed for log storage}
\format{
An object of class \code{character} of length 2.
}
\usage{
AVAILABLE_DB_DRIVERS
}
\description{
These values can be used as DRV parameter inside config_file.
}
\keyword{datasets}
