% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging-tools.R
\docType{data}
\name{browser_info_js}
\alias{browser_info_js}
\alias{log_browser_version}
\title{Browser info}
\format{An object of class \code{html} (inherits from \code{character}) of length 1.}
\usage{
browser_info_js

log_browser_version(input, user_connection_data, user_session, user)
}
\description{
It sends info about user's browser to server.
Place it inside head tag of your Shiny app. You can get this value on server from \code{input[["browser_version"]]}.
You can also use log_browser_version function to log browser version into sqlite file.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
library(shiny)
library(shiny.semantic)
library(shiny.admin)

ui <- function() {
  shinyUI(
    semanticPage(
      tags$head(shiny.admin::browser_info_js),
      title = "Browser info example",
      textOutput("browser")
    )
  )
}

server <- shinyServer(function(input, output) {
  output$browser <- renderText(input[["browser_version"]])
})

shinyApp(ui = ui(), server = server)
}
}
\keyword{datasets}
