% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber-shared.R
\name{build_id_from_secret}
\alias{build_id_from_secret}
\title{Builds id from a secret that can be used in open communication}
\usage{
build_id_from_secret(secret)
}
\arguments{
\item{secret}{string that contains information that should not be publicly
available}
}
\value{
a string with an hash of the secret
}
\description{
This is used in shiny.telemetry, but also externally with the Plumber
endpoint.
}
\examples{
build_id_from_secret("some_random_secret_generated_with_uuid::UUIDgenerate")
}
