% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{build_query_sql}
\alias{build_query_sql}
\title{Build query to read from table in DataStorageSQLite provider}
\usage{
build_query_sql(bucket, date_from = NULL, date_to = NULL)
}
\arguments{
\item{bucket}{string with name of table}

\item{date_from}{date representing the starting day of results. Can be NULL.}

\item{date_to}{date representing the last day of results. Can be NULL.}
}
\value{
string with SQL query
}
\description{
Build query to read from table in DataStorageSQLite provider
}
\examples{
build_query_sql("table_name")
build_query_sql("table_name", Sys.Date() - 365)
build_query_sql("table_name", date_to = Sys.Date() + 365)
build_query_sql("table_name", Sys.Date() - 365, Sys.Date() + 365)
build_query_sql("table_name", as.Date("2023-04-13"), as.Date("2000-01-01"))
}
