% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity-tools.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect with sql database}
\usage{
connect_to_db(db_credentials)
}
\arguments{
\item{config_list}{data.frame with database config parameters}
}
\description{
Makes connection to database based on passed config data
}
