% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber-shared.R
\docType{data}
\name{date_from_null}
\alias{date_from_null}
\title{Common date_from to recognize as NULL}
\format{
An object of class \code{character} of length 1.
}
\usage{
date_from_null
}
\description{
Common date_from to recognize as NULL
}
\keyword{datasets}
