% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber-shared.R
\docType{data}
\name{date_to_null}
\alias{date_to_null}
\title{Common date_to to recognize as NULL}
\format{
An object of class \code{character} of length 1.
}
\usage{
date_to_null
}
\description{
Common date_to to recognize as NULL
}
\keyword{datasets}
