% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty-app.R
\name{empty_app}
\alias{empty_app}
\title{Shiny app to run when no telemetry data is available}
\usage{
empty_app()
}
\value{
An object that represents the empty shiny app. Printing the object or
passing it to \code{shiny::runApp()} will run it.
}
\description{
Shiny app to run when no telemetry data is available
}
