% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity-tools.R
\name{get_users_data}
\alias{get_users_data}
\title{get users data}
\usage{
get_users_data(users_credentials_db_config)
}
\arguments{
\item{users_credentials_db_config}{Named vector with environmental variables concerning db connection.}
}
\description{
get users data
}
