% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity-tools.R
\name{initialize_connection}
\alias{initialize_connection}
\alias{initialize_custom_connection}
\title{Register connection to database}
\usage{
initialize_connection(db_connection)

initialize_custom_connection(db_connection, table_schemes)
}
\arguments{
\item{db_connection}{Connection to database object.}

\item{table_schemes}{Specific table schemes that should be created or connected with inside 'db_connection'.
Each list object specifies separate table and should be of the form:
table_name = c(col_1_name = col_1_type, ...)}
}
\value{
Return list with two objects:
session_id (useful to specify user session after app is opened),
db_connection the same object that was passed as db_connection parameter.

It also sets empty 'user_log' table with:
\code{CREATE TABLE user_log(time TIMESTAMP, session TEXT, username TEXT, action TEXT, id TEXT, value TEXT)}
and 'session_details' table with:
\code{CREATE TABLE session_details(time TIMESTAMP, session TEXT, detail TEXT)} inside DB.
}
\description{
This connects to specified sqlite database (or create new one when it doesn't exist).
Should be used outside server and ui function.
}
