% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MAX_POSTGRES_CONNECTIONS}
\alias{MAX_POSTGRES_CONNECTIONS}
\title{Maximum of connections}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MAX_POSTGRES_CONNECTIONS
}
\description{
Set maximum of connections to postgres from one session
}
\keyword{datasets}
