% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-panel-server.R
\name{prepare_admin_panel_components}
\alias{prepare_admin_panel_components}
\title{prepare_admin_panel_components}
\usage{
prepare_admin_panel_components(input, output, session, data_storage)
}
\arguments{
\item{input}{input object inherited from server function.}

\item{output}{output object inherited from server function.}

\item{session}{session object inherited from server function.}

\item{data_storage}{data_storage instance that will handle all backend read
and writes.}
}
\description{
prepare_admin_panel_components
}
