% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managing-users.R
\name{register_user}
\alias{register_user}
\title{Registers a new user in the database.}
\usage{
register_user(connection, username, password, roles = list(), info = list())
}
\arguments{
\item{connection}{Database connection.}

\item{username}{A string representing unique user name.}

\item{password}{A plain text password that will be encrypted and saved.}

\item{roles}{A list of roles. Each role is a string representing role name.}

\item{info}{A named list containing additional information about the user.}
}
\description{
Registers a new user in the database.
}
