% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managing-users.R
\name{retrieve_users}
\alias{retrieve_users}
\title{Retrieves all users stored in the database.}
\usage{
retrieve_users(connection)
}
\arguments{
\item{connection}{Database connection.}
}
\description{
Retrieves all users stored in the database.
}
