% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-app.R
\name{run_analytics_dashboard}
\alias{run_analytics_dashboard}
\title{Run example telemetry analytics dashboard}
\usage{
run_analytics_dashboard(data_storage)
}
\arguments{
\item{data_storage}{data_storage instance that will handle all backend read
and writes.}
}
\description{
Run example telemetry analytics dashboard
}
