% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-panel-top.R
\name{shiny_stats_ui}
\alias{shiny_stats_ui}
\title{Stats Shiny UI}
\usage{
shiny_stats_ui(custom_css_path = NULL)
}
\arguments{
\item{custom_css_path}{path to custom css}
}
\description{
Stats Shiny UI
}
