% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity-tools.R
\name{source_renviron}
\alias{source_renviron}
\title{Source renviron from file}
\usage{
source_renviron(path, variable_name)
}
\arguments{
\item{path}{path to Renviron file}

\item{variable_name}{name of variable to return with the function}
}
\value{
environmental variable
}
\description{
Source renviron from file
}
