% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_elements.R
\name{use_telemetry}
\alias{use_telemetry}
\title{Function that adds telemetry HTML elements to UI}
\usage{
use_telemetry(id = "")
}
\arguments{
\item{id}{(optional) string with id representing the namespace}
}
\value{
A \code{shiny.tag} object to be included in the UI of a Shiny app.
}
\description{
Function that adds telemetry HTML elements to UI
}
