% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-app.R
\name{analytics_app}
\alias{analytics_app}
\title{Run example telemetry analytics dashboard}
\usage{
analytics_app(data_storage, css_path = NULL)
}
\arguments{
\item{data_storage}{data_storage instance that will handle all backend read
and writes.}

\item{css_path}{string path to css file}
}
\description{
Run example telemetry analytics dashboard
}
