% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-panel-top.R
\name{analytics_app}
\alias{analytics_app}
\title{Run example telemetry analytics dashboard}
\usage{
analytics_app(
  get_user,
  allow_admin_rule = function(session) TRUE,
  db_credentials
)
}
\arguments{
\item{get_user}{test}

\item{allow_admin_rule}{test}

\item{db_credentials}{test}
}
\description{
Run example telemetry analytics dashboard
}
