% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-app.R
\name{analytics_app}
\alias{analytics_app}
\title{Run example telemetry analytics dashboard}
\usage{
analytics_app(data_storage)
}
\arguments{
\item{data_storage}{data_storage instance that will handle all backend read
and writes.}
}
\value{
An object that represents the analytics app. Printing the object or
passing it to `shiny::runApp()` will run it.
}
\description{
Run example telemetry analytics dashboard
}
