% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-storage.R
\name{DataStorage}
\alias{DataStorage}
\title{Data storage abstract class}
\description{
Data storage abstract class

Data storage abstract class
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{username}}{string field that returns the current username}

\item{\code{session_id}}{string field that returns the current session_id}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataStorage-new}{\code{DataStorage$new()}}
\item \href{#method-DataStorage-insert}{\code{DataStorage$insert()}}
\item \href{#method-DataStorage-read_user_data}{\code{DataStorage$read_user_data()}}
\item \href{#method-DataStorage-read_session_data}{\code{DataStorage$read_session_data()}}
\item \href{#method-DataStorage-close}{\code{DataStorage$close()}}
\item \href{#method-DataStorage-clone}{\code{DataStorage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorage-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorage-new}{}}}
\subsection{Method \code{new()}}{
initialize data storage object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorage$new(username, session_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{string with username of the current session}

\item{\code{session_id}}{string with custom session id (should not be used)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorage-insert"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorage-insert}{}}}
\subsection{Method \code{insert()}}{
Insert new data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorage$insert(values, bucket = "user_log")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{values}}{list of values to write to storage provider}

\item{\code{bucket}}{string with name of type of data to write (example, for
SQL it should represent a table)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorage-read_user_data"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorage-read_user_data}{}}}
\subsection{Method \code{read_user_data()}}{
read all user data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorage$read_user_data(date_from, date_to)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date_from}}{date representing the starting day of results}

\item{\code{date_to}}{date representing the last day of results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorage-read_session_data"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorage-read_session_data}{}}}
\subsection{Method \code{read_session_data()}}{
read all session data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorage$read_session_data(date_from, date_to)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date_from}}{date representing the starting day of results}

\item{\code{date_to}}{date representing the last day of results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorage-close"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorage-close}{}}}
\subsection{Method \code{close()}}{
read all session data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorage$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
