% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-storage.R
\name{DataStorage}
\alias{DataStorage}
\title{Data Storage abstract class to handle all the read/write operations}
\description{
Abstract R6 Class that encapsulates all the operations needed by
Shiny.telemetry to read and write. This removes the complexity from the
functions and uses a unified API.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{username}}{string field that returns the current username}

\item{\code{session_id}}{string field that returns the current session_id}

\item{\code{action_bucket}}{string that identifies the bucket to store user
related and action data}

\item{\code{session_bucket}}{string that identifies the bucket to store session
details data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataStorage-new}{\code{DataStorage$new()}}
\item \href{#method-DataStorage-insert}{\code{DataStorage$insert()}}
\item \href{#method-DataStorage-read_user_data}{\code{DataStorage$read_user_data()}}
\item \href{#method-DataStorage-read_session_data}{\code{DataStorage$read_session_data()}}
\item \href{#method-DataStorage-close}{\code{DataStorage$close()}}
\item \href{#method-DataStorage-clone}{\code{DataStorage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorage-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorage-new}{}}}
\subsection{Method \code{new()}}{
initialize data storage object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorage$new(username, session_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{string with username of the current session}

\item{\code{session_id}}{string with custom session id. We recommend using the
session$token from Shiny}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorage-insert"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorage-insert}{}}}
\subsection{Method \code{insert()}}{
Insert new data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorage$insert(
  values,
  bucket = "user_log",
  add_username = TRUE,
  force_params = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{values}}{list of values to write to storage provider}

\item{\code{bucket}}{string with name of type of data to write (example, for
SQL it should represent a table)}

\item{\code{add_username}}{boolean flag that indicates if line should include
the username of the current session}

\item{\code{force_params}}{boolean flag that indicates if parameters should
be generated automatically}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorage-read_user_data"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorage-read_user_data}{}}}
\subsection{Method \code{read_user_data()}}{
read all user data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorage$read_user_data(date_from, date_to)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date_from}}{date representing the starting day of results}

\item{\code{date_to}}{date representing the last day of results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorage-read_session_data"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorage-read_session_data}{}}}
\subsection{Method \code{read_session_data()}}{
read all session data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorage$read_session_data(date_from, date_to)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date_from}}{date representing the starting day of results}

\item{\code{date_to}}{date representing the last day of results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorage-close"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorage-close}{}}}
\subsection{Method \code{close()}}{
read all session data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorage$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
