% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-storage-sqlite.R
\name{DataStorageRSQLite}
\alias{DataStorageRSQLite}
\title{Data storage class with SQLite provider}
\description{
Implementation of the DataStorage R6 class to SQLite backend using a unified
API for read/write operations
}
\examples{
data_storage <- DataStorageRSQLite$new(
  db_path = tempfile(pattern = "user_stats", fileext = ".sqlite")
)
log_login(data_storage)

log_click(data_storage, "an_id")
log_click(data_storage, "a_different_id")

log_session_detail(data_storage, detail = "some detail")

data_storage$read_user_data(as.Date("2020-01-01"), as.Date("2025-01-01"))
data_storage$read_user_data("2020-01-01", "2025-01-01")
}
\section{Super class}{
\code{\link[shiny.telemetry:DataStorage]{shiny.telemetry::DataStorage}} -> \code{DataStorageRSQLite}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataStorageRSQLite-new}{\code{DataStorageRSQLite$new()}}
\item \href{#method-DataStorageRSQLite-insert}{\code{DataStorageRSQLite$insert()}}
\item \href{#method-DataStorageRSQLite-read_user_data}{\code{DataStorageRSQLite$read_user_data()}}
\item \href{#method-DataStorageRSQLite-read_session_data}{\code{DataStorageRSQLite$read_session_data()}}
\item \href{#method-DataStorageRSQLite-close}{\code{DataStorageRSQLite$close()}}
\item \href{#method-DataStorageRSQLite-clone}{\code{DataStorageRSQLite$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorageRSQLite-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorageRSQLite-new}{}}}
\subsection{Method \code{new()}}{
Initialize the data storage class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorageRSQLite$new(db_path = "user_stats.sqlite")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db_path}}{string with path to sqlfile}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorageRSQLite-insert"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorageRSQLite-insert}{}}}
\subsection{Method \code{insert()}}{
Insert new data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorageRSQLite$insert(values, bucket = self$action_bucket)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{values}}{list of values to write to database}

\item{\code{bucket}}{name of table to write}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorageRSQLite-read_user_data"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorageRSQLite-read_user_data}{}}}
\subsection{Method \code{read_user_data()}}{
read all user data from SQLite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorageRSQLite$read_user_data(date_from, date_to)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date_from}}{date representing the starting day of results}

\item{\code{date_to}}{date representing the last day of results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorageRSQLite-read_session_data"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorageRSQLite-read_session_data}{}}}
\subsection{Method \code{read_session_data()}}{
read all session data from SQLite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorageRSQLite$read_session_data(date_from, date_to)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date_from}}{date representing the starting day of results}

\item{\code{date_to}}{date representing the last day of results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorageRSQLite-close"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorageRSQLite-close}{}}}
\subsection{Method \code{close()}}{
read all session data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorageRSQLite$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorageRSQLite-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorageRSQLite-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorageRSQLite$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
