% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-storage-sqlite.R
\name{DataStorageSQLite}
\alias{DataStorageSQLite}
\title{Data storage class with SQLite provider}
\description{
Implementation of the DataStorage R6 class to SQLite backend using a unified
API for read/write operations
}
\examples{
data_storage <- DataStorageSQLite$new(
  db_path = tempfile(pattern = "user_stats", fileext = ".sqlite")
)

telemetry <- Telemetry$new(data_storage = data_storage)
telemetry$log_login()

telemetry$log_click("an_id")
telemetry$log_click("a_different_id")

data_storage$read_event_data("2020-01-01", "2025-01-01")
}
\section{Super class}{
\code{\link[shiny.telemetry:DataStorage]{shiny.telemetry::DataStorage}} -> \code{DataStorageSQLite}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataStorageSQLite-new}{\code{DataStorageSQLite$new()}}
\item \href{#method-DataStorageSQLite-read_event_data}{\code{DataStorageSQLite$read_event_data()}}
\item \href{#method-DataStorageSQLite-clone}{\code{DataStorageSQLite$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="shiny.telemetry" data-topic="DataStorage" data-id="close"><a href='../../shiny.telemetry/html/DataStorage.html#method-DataStorage-close'><code>shiny.telemetry::DataStorage$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny.telemetry" data-topic="DataStorage" data-id="insert"><a href='../../shiny.telemetry/html/DataStorage.html#method-DataStorage-insert'><code>shiny.telemetry::DataStorage$insert()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorageSQLite-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorageSQLite-new}{}}}
\subsection{Method \code{new()}}{
Initialize the data storage class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorageSQLite$new(db_path = "user_stats.sqlite")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db_path}}{string with path to sqlfile}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorageSQLite-read_event_data"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorageSQLite-read_event_data}{}}}
\subsection{Method \code{read_event_data()}}{
read all user data from SQLite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorageSQLite$read_event_data(date_from, date_to)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date_from}}{date representing the starting day of results}

\item{\code{date_to}}{date representing the last day of results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataStorageSQLite-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataStorageSQLite-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataStorageSQLite$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
